RABID-524 select x86 flavour based on CONFIG_X86_*, not ARCH
When building kernel modules the value of ARCH should match the name of
an arch/ subdirectory in the kernel source tree. Thus when building for
x86 one should use ARCH=x86, under which the 32/64-bitness is selected
via the mutually exclusive CONFIG_X86_32 and CONFIG_X86_64 definitions.

This allows passing the standard ARCH=x86 when building for x86-64.
Previously ARCH=x86 would have led, via the symlink, to use the i386
code, which then failed to compile due to the architectural mismatch.

With this change, builders are required to pass ARCH=x86; ARCH=i386 and
ARCH=x86_64 will no longer work.