pass ethernet header to netfilter hook
previously, the `skb` argument of the netfilter hook pointed to the beginning of
the IP header. with this change, in standard cases, it now points to the beginning
of the Ethernet header. In exceptional cases where the Ethernet header is not
present, such as when the hook is attached to `LOCAL_OUT`, the skb object points
to the beginning of the IP header.